requirements.txt
openai
langchain
langchain-community
python-dotenv
streamlit
langchain_openai